requirements.txt
Flask
pandas
openpyxl
python-pptx
gunicorn
